Update th.json
Dear Lluis,

I receive this message when making edtis to this file:

"We’ve detected the file has mixed line endings. When you commit changes we will normalize them to Windows-style (CRLF)."

And it won't allow a preview of the changes. I am not sure what is meant with "normalize" but I hope it wont change the Thai symbols. Please check before submitting the change.